add a property for controlling perf_event_paranoid
This allows the system user to set perf_event_paranoid to either 1 or 2,
as a step towards eliminating the perf events attack surface by default
and having developers opt-in to enabling it. It does not allow reducing
it from the default of perf_event_paranoid=1 to avoid introducing any
new system->root attack surface.

Change-Id: I6e3ae3cf18d8c76df94f879c34fb6fde519b89a9